refactor(store): move record types out of store.ts into store/records/*
Relocates the 43 interfaces, 17 type aliases, ~20 constants and 8
tightly-coupled helper functions (withActiveDay, applyMeasuredTokens,
isStudioOrigin, dispatchAttempt, telemetryExpiresAt,
gameAssessmentId/hydrateGameAssessment/reReviewRequestId,
compareProposals) that made up store.ts's record-types region into
apps/api/src/store/records/<slice>.ts, one file per domain in the
Phase 2 store-split taxonomy (identity, rounds, submission, dispatch,
build-log, quota, telemetry, notifications, social, review,
player-data, contribution, access, access-tokens, agent-keys, oauth).

store.ts re-exports every moved symbol under its original name, so
none of the 183 existing importers change in this commit -- each
slice's own consumers migrate to the direct import path when that
slice is carved out of the Store interface in a later wave.

store.ts: 8833 -> 7060 lines (record region was ~1800 lines; some
shrank further via prettier's tighter import wrapping). Ceiling
resealed down to match. New record files get their own module-size
and comment-prose baselines sealed at their actual (relocated, not
new) content -- this is moved prose, not new debt.

Phase 2 wave 1 of the store-split plan (docs/phase2-store-split-
execution-plan.md in the private ops repo).
