+++ eslint-rules/comment-prose-baseline.json
@@ -274,6 +274,20 @@
     "apps/api/src/store-firestore.test.ts": 566,
     "apps/api/src/store.test.ts": 486,
     "apps/api/src/store.ts": 16323,
+    "apps/api/src/store/records/access-tokens.ts": 204,
+    "apps/api/src/store/records/agent-keys.ts": 169,
+    "apps/api/src/store/records/build-log.ts": 492,
+    "apps/api/src/store/records/contribution.ts": 1107,
+    "apps/api/src/store/records/dispatch.ts": 598,
+    "apps/api/src/store/records/identity.ts": 376,
+    "apps/api/src/store/records/notifications.ts": 618,
+    "apps/api/src/store/records/oauth.ts": 21,
+    "apps/api/src/store/records/player-data.ts": 710,
+    "apps/api/src/store/records/quota.ts": 308,
+    "apps/api/src/store/records/review.ts": 488,
+    "apps/api/src/store/records/social.ts": 232,
+    "apps/api/src/store/records/submission.ts": 1562,
+    "apps/api/src/store/records/telemetry.ts": 1136,
     "apps/api/src/submission-status.test.ts": 35,
     "apps/api/src/submission-status.ts": 257,
     "apps/api/src/submissions.test.ts": 4563,
